+++ lastEnd/untitled/src/main/java/com/heavenfilms/backend/repository/CinemaRepository.java
@@ -3,5 +3,5 @@ package com.heavenfilms.backend.repository;
 import com.heavenfilms.backend.entity.Cinema;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CinemaRepository extends JpaRepository<Cinema, Integer> {
+public interface CinemaRepository extends JpaRepository<Cinema, Long> {
 }
