+++ lastEnd/untitled/src/main/java/com/heavenfilms/backend/repository/FavoriteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
-    Optional<Favorite> findByUserIdAndMovieId(Integer userId, Integer movieId);
-    List<Favorite> findByUserId(Integer userId);
+public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+    Optional<Favorite> findByUserIdAndMovieId(Long userId, Long movieId);
+    List<Favorite> findByUserId(Long userId);
 }
