+++ lastEnd/untitled/src/main/java/com/heavenfilms/backend/service/MovieService.java
@@ -2,7 +2,6 @@ package com.heavenfilms.backend.service;
 
 import com.heavenfilms.backend.entity.Movie;
 import org.springframework.data.domain.Page;
-
 import java.util.List;
 
 public interface MovieService {
@@ -10,7 +9,7 @@ public interface MovieService {
     Page<Movie> getHotMovies(String genre, int page, int size);
     Page<Movie> getUpcomingMovies(String genre, int page, int size);
     Page<Movie> getMoviesByTag(String tag, int page, int size);
-    Movie getMovieById(Integer id);
+    Movie getMovieById(Long id);
     List<Movie> getAllMovies();
-    List<Movie> getMoviesByIds(List<Integer> ids);
+    List<Movie> getMoviesByIds(List<Long> ids);
 }
