+++ lastEnd/untitled/src/main/java/com/heavenfilms/backend/service/ScheduleService.java
@@ -1,11 +1,10 @@
 package com.heavenfilms.backend.service;
 
 import com.heavenfilms.backend.entity.Schedule;
-
 import java.util.List;
 import java.util.Map;
 
 public interface ScheduleService {
-    List<Map<String, Object>> getSchedulesByMovie(Integer movieId);
-    Schedule getScheduleById(Integer id);
+    List<Map<String, Object>> getSchedulesByMovie(Long movieId);
+    Schedule getScheduleById(Long id);
 }
