+++ lastEnd/untitled/src/main/java/com/heavenfilms/backend/service/UserService.java
@@ -7,5 +7,5 @@ public interface UserService {
     UserDTO login(String account, String password);
     UserDTO register(String account, String password, String nickname, String phone);
     UserDTO update(User user);
-    UserDTO findById(Integer id);
+    UserDTO findById(Long id);
 }
